Add files via upload
Peter Allwin's original 1.6 Plaits port code including the noisy models too compiled for all three platforms. still missing Modal, which has a large compile bug, and Speech, this needs to be cut up like that wave table oscillator as much of it is speech sample waveforms with static definitions. some are redundant with Peter's models, but it was just easier to do them all at once. this is a baseline to get 1.2 running.